Assignment 4

+++ user-management/src/main/java/org/ac/cst8277/kim/riyoun/usermanagement/Controller.java
@@ -1,21 +1,24 @@
 package org.ac.cst8277.kim.riyoun.usermanagement;
 
 import java.sql.Timestamp;
-import java.util.ArrayList;
-import java.util.List;
+import java.time.Duration;
+import java.time.Instant;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-import java.util.function.Function;
-import java.time.Instant;
 
-import org.springframework.web.bind.annotation.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.oauth2.core.user.OAuth2User;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 import lombok.Data;
 
@@ -27,74 +30,46 @@ public class Controller {
     @Autowired
     private RoleRepository roleRepo;
 
-    @Autowired
-    private SessionRepository sessionRepo;
-
-    @GetMapping("/github")
+    @GetMapping("/user")
     public Map<String, Object> user(@AuthenticationPrincipal OAuth2User principal) {
         return principal.getAttributes();
     }
 
     @GetMapping("/user/all")
     public ResponseEntity<?> getAllUsers(@AuthenticationPrincipal OAuth2User principal) {
-        List<String> roles = new ArrayList<String>();
-        for (GrantedAuthority r : principal.getAuthorities()) {
-            roles.add(r.getAuthority());
-        };
-        if (roles.contains("ADMIN")) {
-            return ResponseEntity.ok().body(userRepo.findAll());
-        }
-        return ResponseEntity
-                    .status(HttpStatus.UNAUTHORIZED)
-                    .body("You have to be admin to do this.");
-    }
-
-    @GetMapping("/user/{id}")
-    public ResponseEntity<?> getUserWithId(@PathVariable("id") String id, @AuthenticationPrincipal OAuth2User principal) {
-        List<String> roles = new ArrayList<String>();
-        for (GrantedAuthority r : principal.getAuthorities()) {
-            roles.add(r.getAuthority());
-        };
-        if (roles.contains("ADMIN")) {
-            UUID uuid = UUID.fromString(id);
-            Optional<User> user = userRepo.findById(uuid);
-            if (user.isEmpty()) {
-                return ResponseEntity
-                    .status(HttpStatus.NOT_FOUND)
-                    .body("No user found. To create a new account, POST to '/usm/user' with username and password.");
-            } else {
-                return ResponseEntity.ok().body(user.get());
-            }
-        }
-        return ResponseEntity
-                    .status(HttpStatus.UNAUTHORIZED)
-                    .body("You have to be admin to do this.");
-    }
-
-    @GetMapping("/user")
-    public ResponseEntity<?> loginUser(@RequestParam String username, @RequestParam String password) {
-        Optional<User> user = userRepo.findByUsernameAndPassword(username, password);
-        if (user.isEmpty()) {
+        if (!hasAuthority(principal, "ADMIN")) {
             return ResponseEntity
                     .status(HttpStatus.UNAUTHORIZED)
-                    .body("No user found. To create a new account, POST to '/usm/user' with username and password.");
+                    .body("You don't have permission to do this.");
+        }
+        return ResponseEntity.ok().body(userRepo.findAll());
+    }
+
+    @GetMapping("/user/{name}")
+    public ResponseEntity<?> getUserWithId(@PathVariable("name") String name, @AuthenticationPrincipal OAuth2User principal) {
+        if (!hasAuthority(principal, "ADMIN") && !principal.getName().equals(name)) {
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body("You don't have permission to do this.");
+        }
+        Optional<User> user = userRepo.findByUsername(name);
+        if (user.isEmpty()) {
+            return ResponseEntity
+                .status(HttpStatus.NOT_FOUND)
+                .body("No user found. To create a new account, POST to '/ums/user' with username and password.");
         } else {
             return ResponseEntity.ok().body(user.get());
         }
     }
 
     @GetMapping("/role/all")
-    public ResponseEntity<?> getAllRoles(@RequestParam String id) {
-        return authenticate(id, user -> {
-            for (Role r : user.getRoles()) {
-                if (r.getRole().equals("ADMIN")) {
-                    return ResponseEntity.ok().body(roleRepo.findAll());
-                }
-            }
+    public ResponseEntity<?> getAllRoles(@AuthenticationPrincipal OAuth2User principal) {
+        if (!hasAuthority(principal, "ADMIN")) {
             return ResponseEntity
-                .status(HttpStatus.UNAUTHORIZED)
-                .body("Your are not authorized to do this.");
-        });
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body("You don't have permission to do this.");
+        }
+        return ResponseEntity.ok().body(roleRepo.findAll());
     }
 
     @PostMapping("/user")
@@ -103,38 +78,33 @@ public class Controller {
         Session session = new Session();
 
         user.setUsername(request.getUsername());
-        user.setPassword(request.getPassword());
-        session.setLastLogin(Timestamp.from(Instant.now()));
-        session.setLastLogout(Timestamp.from(Instant.now()));
+        session.setLogin(Timestamp.from(Instant.now()));
+        session.setExpire(Timestamp.from(Instant.now().plus(Duration.ofMinutes(15))));
         user.setLastSession(session);
 
         User u = userRepo.save(user);
-        sessionRepo.save(session);
         return u;
     }
 
-    @DeleteMapping("/users/{id}")
-    public void deleteById(@PathVariable("id") String id) {
+    @DeleteMapping("/user/{id}")
+    public ResponseEntity<?> deleteById(@PathVariable("id") String id, @AuthenticationPrincipal OAuth2User principal) {
+        if (!hasAuthority(principal, "ADMIN")) {
+            return ResponseEntity
+                    .status(HttpStatus.UNAUTHORIZED)
+                    .body("You don't have permission to do this.");
+        }
         UUID uuid = UUID.fromString(id);
         userRepo.deleteById(uuid);
+        return ResponseEntity.ok().body("Deleted user with ID " + id);
     }
 
-    private ResponseEntity<?> authenticate(String id, Function<User, ResponseEntity<?>> callback) {
-        try {
-            UUID uuid = UUID.fromString(id);
-            Optional<User> user = userRepo.findById(uuid);
-            if (user.isEmpty()) {
-                return ResponseEntity
-                    .status(HttpStatus.UNAUTHORIZED)
-                    .body("No user found. To create a new account, POST to '/usm/user' with username and password.");
-            } else {
-                return callback.apply(user.get());
-            }
-        } catch (IllegalArgumentException e) {
-            return ResponseEntity
-                .badRequest()
-                .body("Illegal UUID. To create a new account, POST to '/usm/user' with username and password.");
+    private boolean hasAuthority(OAuth2User principal, String authority) {
+        System.out.println(principal.getName());
+        for (GrantedAuthority r : principal.getAuthorities()) {
+            System.out.println(r.getAuthority());
+            if (authority.equals(r.getAuthority())) return true;
         }
+        return false;
     }
 }
